Addition of dialogue system
Dialog scripts added. Dialog works when  on-screen button in test scene is clicked, but the dialog trigger on the sign object is broken. It only displays its text on the text box when the button has been clicked beforehand, and does not close the text box upon a second trigger like it should.